LIBDRUM-650. Override the overlays module in quick build.
The overlay modules 'additions' and 'server' should be deleted from the base image before current the files added to them for the quick build. This is necessary to ensure newly deleted files in the overlays module are indeed not included (inherited from the base build) while doing the quick builds.

https://issues.umd.edu/browse/LIBDRUM-650