LIBDRUM-763. Customized "OpenSearchServiceImpl" to reset server url
In DRUM, when running in production, we run both the front-end and
back-end at the same URL, using the Kubernetes Ingress and rewrite rules
to send requests to either Angular or the DSpace server, as appropriate.

For OpenSearch, when run in Kubernetes, we want to use the Angular URL,
instead of the DSpace server URL, in the OpenSearch output (while
retaining the ability to use separate servers in local development).

To enable this, customized the "getBaseSearchServiceURL()" method in
the "OpenSearchServiceImpl" class to look for a UMD custom
"opensearch.server.url" property (which enables us to customize the
server hostname in Kubernetes), which defaults to the stock
"dspace.server.url" property if an "opensearch.server.url" property
value is not present.

https://umd-dit.atlassian.net/browse/LIBDRUM-763