LIBDRUM-991. Limit file size process by EtdLoader
In Kubernetes, EtdLoader uncompresses files into an ephemeral
volume with limited space. If an uncompressed file exceeds
the ephemeral volume size, the pod will reboot.

To prevent this, added a “drum.etdloader.maxFileSize”
configuration property. A Zip file containing a ZipEntry that
is larger than this size will not be processed, and a
prominent error message will be added to the log.

Updated EtdLoaderTest with a test of the “drum.etdloader.maxFileSize” property and error message.

https://umd-dit.atlassian.net/browse/LIBDRUM-991