LIBDRUM-991. Report errors to enclosing script
Modified EtdLoader to exit with a non-zero status code when an error
occurs in the processing.

Modified the "load-etd" script to capture the status code from the
EtdLoader and use it as its exit status.

Modified the "load-etd-nightly" script to print an error message and
continue to the next file (leaving the file with the error in the
"incoming" directory, instead of moving it to the "processed"
directory), and return an exit status.

Modified "script-main-wrapper" to capture exit status and modify
the email "Subject" line to include "- ERROR(S) OCCURRE" on a non-zero
exit status.

https://umd-dit.atlassian.net/browse/LIBDRUM-991